pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ceds-jsonld"
version = "0.1.0"
description = "Python library for ingesting education data, mapping to CEDS/CEPI ontology SHACL shapes, and outputting conformant JSON-LD"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    { name = "CEPI", email = "cepi@michigan.gov" },
]
dependencies = [
    "pyyaml>=6.0",
    "rdflib>=7.0",
    "pandas>=2.0",
]

[project.optional-dependencies]
fast = ["orjson>=3.10"]
excel = ["openpyxl>=3.1"]
cosmos = ["azure-cosmos>=4.7", "azure-identity>=1.15"]
validation = ["pyshacl>=0.26"]
api = ["httpx>=0.27"]
database = ["sqlalchemy>=2.0"]
cli = ["click>=8.1"]
observability = ["structlog>=24.0", "tqdm>=4.60"]
all = [
    "orjson>=3.10",
    "openpyxl>=3.1",
    "azure-cosmos>=4.7",
    "azure-identity>=1.15",
    "pyshacl>=0.26",
    "httpx>=0.27",
    "sqlalchemy>=2.0",
    "click>=8.1",
    "structlog>=24.0",
    "tqdm>=4.60",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-httpserver>=1.0",
    "hypothesis>=6.0",
    "ruff>=0.5",
    "mypy>=1.10",
    "orjson>=3.10",
    "pyshacl>=0.26",
    "openpyxl>=3.1",
    "httpx>=0.27",
    "sqlalchemy>=2.0",
    "structlog>=24.0",
    "tqdm>=4.60",
    "azure-cosmos>=4.7",
    "azure-identity>=1.15",
    "click>=8.1",
    "sphinx>=7.0",
    "sphinx-rtd-theme>=2.0",
    "sphinx-autodoc-typehints>=2.0",
]

[project.scripts]
ceds-jsonld = "ceds_jsonld.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "benchmark: marks tests as performance benchmarks",
    "slow: marks tests as slow",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.hatch.build.targets.wheel]
packages = ["src/ceds_jsonld"]

[tool.coverage.run]
source = ["src/ceds_jsonld"]
omit = ["tests/*"]
